refactor(umbp): share host registration and gather helpers
